server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["vikyw89 <vikyw89@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
llama-index = "^0.9.11"
google-cloud-aiplatform = "^1.36.4"
python-dotenv = "^1.0.0"
prisma = "^0.11.0"
pytest = "^7.4.3"
httpx = "^0.24.0"
langchain = "^0.0.346"
redis = "^5.0.1"
supabase = "^2.2.0"
python-magic = "^0.4.27"
firebase-admin = "^6.3.0"
google-cloud-storage = "^2.13.0"
cloudinary = "^1.37.0"
google-cloud-speech = "^2.23.0"
google-cloud-texttospeech = "^2.15.0"
google = "^3.0.0"
python-multipart = "^0.0.6"
ffmpeg = "^1.4"
ffprobe = "^0.5"
pydub = "^0.25.1"
moviepy = "^1.0.3"
tinytag = "^1.10.1"
google-auth = "^2.25.2"
google-auth-httplib2 = "^0.1.1"
google-auth-oauthlib = "^1.1.0"
ffmpeg-python = "^0.2.0"
pyo = "^1.0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
